+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/EntityMixin.java
@@ -3,12 +3,11 @@ package insane96mcp.iguanatweaksreborn.mixin;
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import insane96mcp.iguanatweaksreborn.module.client.Misc;
 import insane96mcp.iguanatweaksreborn.module.misc.tweaks.Tweaks;
+import insane96mcp.iguanatweaksreborn.module.movement.BetterClimbable;
 import insane96mcp.iguanatweaksreborn.module.world.Fluids;
 import it.unimi.dsi.fastutil.objects.Object2DoubleMap;
 import net.minecraft.world.damagesource.DamageSource;
-import net.minecraft.world.damagesource.DamageSources;
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.level.Level;
 import net.minecraftforge.common.ForgeMod;
 import net.minecraftforge.fluids.FluidType;
 import org.spongepowered.asm.mixin.Mixin;
@@ -24,17 +23,7 @@ public abstract class EntityMixin {
     @Shadow(remap = false)
     protected Object2DoubleMap<FluidType> forgeFluidTypeHeight;
 
-    @Shadow public abstract float getBbHeight();
-
-    @Shadow public boolean horizontalCollision;
-
-    @Shadow public boolean minorHorizontalCollision;
-
-    @Shadow private Level level;
-
     @Shadow public abstract boolean hurt(DamageSource pSource, float pAmount);
-
-    @Shadow public abstract DamageSources damageSources();
 
     @Inject(at = @At(value = "RETURN"), method = "fireImmune", cancellable = true)
     private void onFireImmune(CallbackInfoReturnable<Boolean> cir) {
@@ -49,7 +38,8 @@ public abstract class EntityMixin {
 
     @Redirect(method = "move", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;resetFallDistance()V"))
     private void onResetFallDistanceInMove(Entity instance) {
-        if (Fluids.shouldOverrideWaterFallDamageModifier())
+        if (Fluids.shouldOverrideWaterFallDamageModifier()
+                || BetterClimbable.fallDamageOnClimbable)
             return;
         this.resetFallDistance();
     }